requirements.txt
folium==0.14.0
inflection==0.5.1
pandas==2.1.1
Pillow==10.1.0
streamlit==1.27.2
streamlit-folium==0.15.0
